Resolve single inspections
Necessary because when the client has not yet seen the seen the single version for
an inspection it doesn't know how to resolve it.

https://formidable.com/open-source/urql/docs/graphcache/local-resolvers/
Given that Graphcache stores entities in a normalized data structure there may be multiple fields on a given schema that can be used to get to the same entity. For instance, the schema may allow for the same entity to be looked up by an ID while this entity may also appear somewhere else in a list or on an entirely different field.

When links (or relations) like these are cached by Graphcache it is able to look up the entities automatically, e.g. if we've sent a { todo(id: 1) { id } } query to our API once then Graphcache will have seen that this field leads to the entity it returns and can query it automatically from its cache.

However, if we have a list like { todos { id } } we may have seen and cached a specific entity, but as we browse the app and query for { todo(id: 1) { id } }, Graphcache isn't able to automatically find this entity even if it has cached it already and will send a request to our API.

In many cases we can create a local resolvers to instead tell the cache where to look for a specific entity by returning partial information for it.