Fix bug where if you first start the exchanges with server offline, it will never show optimistic results
Also use a cache-only requestPolicy to prevent refetching queries every time. It seems that cache-first
gets upgraded to cache-and-network by graphcache upon receiving a mutation result. It's not known why
it decides to do this right now.

This was caused by the dedup exchange filtering out the reexecuted queries, believe them to
be extraneous. It would never see the OperationResult coming back since server offline causes uncaught
errors to be thrown from the fetch exchange. That in turn causes no errored results to come back.

The easy fix is to simply place queries on the rebound channel, rather than reexecuting them
through the client. Whether or not this is "correct" is hard to say.

The underlying fix should be tracked via https://github.com/FormidableLabs/urql/issues/2234