Persist incomplete mutations that pass into the crdtExchange.
Persist the mutations that pass into the crdtExchange so that we
don't lose user data if the client is terminated before the mutation has been
successfully executed by the graphql service.

When the client is restarted, rehydrate the crdtExchange from persisted mutation
data. Buffer new CRDT mutations from entering the crdtExchange until rehydration
is complete.

Additionally, in order to facilitate better testability of the crdtExchange,
remove the hard-coded `interval` Wonka source which triggers sending of crdt
mutations from the CrdtManager. Inject this dependency into the class. This way,
tests can have better control of triggering transmission from the crdtExchange.
In addition, this should make it easier in the future to build a more
sophisticated mechanism to control when the crdtExchange should transmit.

Co-Authored-By: Malcolm Scruggs <malcolm.scruggs@appfolio.com>
Co-Authored-By: Jason Ho <jason.ho@appfolio.com>